pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-fortress"
version = "0.1.0"
description = "Production-grade security firewall & proxy for AI Agent MCP (Model Context Protocol)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "Avoceous", email = "Avoceous@users.noreply.github.com" }
]
keywords = [
    "mcp", "security", "ai-agent", "firewall", "prompt-injection",
    "tool-poisoning", "llm-security", "ai-security", "mcp-fortress",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Zero required dependencies for the core engine
dependencies = []

[project.optional-dependencies]
proxy = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "websockets>=12",
]
pii = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
yaml = [
    "pyyaml>=6.0",
]
dashboard = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
]
all = [
    "mcp-fortress[proxy,pii,yaml,dashboard]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-fortress       = "mcpshield.cli.main:main"
mcp-fortress-stdio = "mcpshield.transport.stdio_proxy:main_stdio"

[project.urls]
Homepage      = "https://github.com/Avoceous/mcp-fortress"
Repository    = "https://github.com/Avoceous/mcp-fortress"
Documentation = "https://github.com/Avoceous/mcp-fortress/blob/main/README.md"
"Bug Tracker" = "https://github.com/Avoceous/mcp-fortress/issues"

[tool.hatch.build.targets.wheel]
packages = ["mcpshield"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B"]
ignore = ["S101", "S603", "S607"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
